+++ src/commands/generator/spring/springBootGenerator.ts
@@ -8,7 +8,7 @@ import { GitHubProjectPersister } from "@atomist/automation-client/operations/ge
 import { cleanReadMe } from "@atomist/automation-client/operations/generate/UniversalSeed";
 import { curry } from "@typed/curry";
 import { addSpringBootStarter } from "../../editor/spring/addStarterEditor";
-import { doUpdatePom, inferStructureAndMovePackage, removeTravisBuildFiles } from "../java/JavaProjectParameters";
+import { cleanTravisBuildFiles, doUpdatePom, inferStructureAndMovePackage } from "../java/JavaProjectParameters";
 import { inferSpringStructureAndRename, SpringBootGeneratorParameters } from "./SpringBootProjectParameters";
 
 export function springBootGenerator(projectPersister: ProjectPersister = GitHubProjectPersister): HandleCommand<SpringBootGeneratorParameters> {
@@ -26,12 +26,12 @@ export function springBootGenerator(projectPersister: ProjectPersister = GitHubP
 export function springBootProjectEditor(params: SpringBootGeneratorParameters): AnyProjectEditor {
     const starterEditors: AnyProjectEditor[] =
         params.starters.map(starter =>
-                addSpringBootStarter("spring-boot-starter-" + starter));
+            addSpringBootStarter("spring-boot-starter-" + starter));
     logger.debug("Starters: [%s]. Editor count=%d", params.starters.join(), starterEditors.length);
 
     const editors: AnyProjectEditor[] = [
         curry(cleanReadMe)(params.target.description),
-        removeTravisBuildFiles,
+        curry(cleanTravisBuildFiles)(slackTeamTravisWebhookUrl(params.slackTeam)),
         curry(doUpdatePom)(params),
         curry(inferStructureAndMovePackage)(params.rootPackage),
         curry(inferSpringStructureAndRename)(params.serviceClassName),
@@ -39,4 +39,8 @@ export function springBootProjectEditor(params: SpringBootGeneratorParameters): 
     return chainEditors(
         ...editors.concat(starterEditors),
     );
+}
+
+function slackTeamTravisWebhookUrl(teamId: string): string {
+    return `https://webhook.atomist.com/atomist/travis/teams/${teamId}`;
 }